feat: add GenesisBuilderMock
This commit adds a builder for initializing a set of N mintnodes, a set
of Y spentbooks, and generating a genesis dbc with amount Z.

This functionality is need in tests, mint-repl, and benches. As such,
it cannot be in tests, but must be part of sn_dbc library.  It is
planned that a future commit will refactor all the "mock" items into
a mock module, behind a feature flag.

* impl builder::mock::GenesisBuilderMock
* use GenesisBuilderMock in mint-repl mk_new-mint()
* use GenesisBuilderMock in lib::tests::issue_genesis()
* improve how rng parameters are passed