style: disambiguate KeyImage usage
The KeyImage wrapper for G1Affine was being used in a few places
where a PublicKeyBlst is used as map key.  But it is semantically
wrong to call it a key image.

This commit changes the wrapper name to PublicKeyBlstMappable and
makes KeyImage a type alias of it.  So now the appropriate name
is used according to context.