style: remove extra validations in dbc builder
This commit removes several validations in DbcBuilder::build().

The only remaining errors returns from ::build() itself are for cases
where an unwrap would otherwise be necessary.

The rationale for remaining these error checks is that:

1) the SpentBook nodes and Mint nodes will validate the transaction
anyway, so the checks are largely redundant.

2) the builder has an incomplete view of the tx because it is
missing the public commitments, so it cannot call
RingCtTransaction::verify() and thus any checks it does will always
be incomplete/subset.

3) It is better to have Tx verification logic written in only one
place.

A counter-argument would be that having the checks enable the client
to easily detect certain errors before contacting spentbook or
mintnode(s).

A counter-counter-argument is that these kind of errors should be
super rare in a properly functioning system, unless some kind of
attack is happening perhaps.

Specifics:

* remove Error::ReissueRequestPublicKeySetMismatch
* remove Error::ReissueRequestPublicCommitmentMismatch
* remove Error::NoReissueShares
* remove Error::ReissueShareMintNodeSignaturesLenMismatch
* remove Error::ReissueShareMintNodeSignatureNotFoundForInput