test: improve reissue_auto in mint-repl
1. configurable reissue_auto
We prompt the user for min and max number of inputs and max outputs
when performing the reissue_auto algo to generate DBCs automatically.

This enables the user to simulate various sizes of reissues and
observe how long they take.

2. timings
Displays some timing information for spentbook operations, wallet
operations, and total time -- for each reissue.