feat: provide hex serialization utilities
A couple of functions are added to serialize and deserialize a `Dbc` to and from a hex encoded
string.

At the moment `sn_cli` needs to perform these conversions, but we decided it would be useful to have
them on the `Dbc` type.

There is also a little change here to switch off verbose output for Cargo on the PR workflow. There
really isn't a lot of useful information from the build process and it clutters the logs.